fix: clear the remaining clang-tidy findings in src/ and tests/
Takes first-party findings to zero. Naming, C-style arrays, mixed-sign
comparisons and narrowing conversions are fixed at the root; the 21
algorithm cores that exceed the cognitive-complexity threshold are
annotated instead of split.

Root-cause fixes rather than casts-in-place:
- ARstart/ARindex/ARvalue collapse onto ar_start/ar_index/ar_value, the
  spelling PropEngine's own constructor and accessors already use.
- The three mixed-sign comparisons on hot paths are removed, not wrapped:
  IndexedMinHeap::contains compares as size_t after its own `var >= 0`
  guard, and LocalMIP's positive_list bound is hoisted out of the loop —
  it is invariant, since compaction never resizes. std::cmp_* is used only
  on the eleven cold sites, where its sign test costs nothing.
- 4 * ncol_ is widened before the multiply, not after.

readability-function-cognitive-complexity stays enabled at threshold 25.
Each of the 21 over-threshold functions carries its measured complexity and
the reason it is one function, per the standards' guidance for genuinely
complex domains: they are the algorithm cores (propagate 92, infeasible_step
93, LocalMipWorker::run_attempt 132) and splitting them moves work across a
worker's inner loop, which cannot be validated without a benchmark. A global
disable would have let the next 200-complexity function through silently.

E, R and Q in fpr_core/repair_search keep their case: they are the paper's
Fig. 5 symbols, used under those names in the surrounding header prose and
as repair_search()'s own parameter.

+++ src/prop_engine.cpp
@@ -272,6 +272,11 @@ void PropEngine::seed_worklist(HighsInt j) {
     }
 }
 
+// Cognitive complexity 92 (threshold 25).  Kept whole: the AC-3 bound-propagation fixpoint; the
+// branch structure is the row/column activity case analysis and it runs per DFS node. Decomposing
+// it would move work across a worker's inner loop, and the closeout takes no unmeasured performance
+// risk; the standards also rank fidelity to the reference algorithm above mechanical extraction.
+// NOLINTNEXTLINE(readability-function-cognitive-complexity)
 bool PropEngine::propagate(HighsInt fixed_var) {
     // Hoist raw pointers once so the tight loop doesn't re-load member
     // fields on every iteration. All of these arrays are read-only (the
@@ -616,7 +621,7 @@ void PropEngine::update_activities(HighsInt j, const VarState& old_vs) {
 void PropEngine::init_domain_pq() {
     domain_pq_.clear();
     pq_undo_.clear();
-    pq_undo_.reserve(4 * ncol_);
+    pq_undo_.reserve(4 * static_cast<size_t>(ncol_));
     for (HighsInt j = 0; j < ncol_; ++j) {
         if (is_int(j) && !vs_[j].fixed) {
             domain_pq_.insert(vs_[j].ub - vs_[j].lb, j);
